feat(worker): move statement state machine into kernel with pure tests
Extract selectPendingQueueNames and ALLOWED_WORKER_QUEUES parsing so fan-out behavior is covered without Mongo, and keep worker modules as re-exports.
